+++ BackEnd/src/main/java/com/example/FinalProject/productCheckOnThePage/NameChecking.java
@@ -18,25 +18,9 @@ public class NameChecking {
 
 
     public List<String> checkingName(String pageSelector){
-        List<String> elements = driver.findElements(By.cssSelector("["+pageSelector+"]"))
-                .stream()
-                .map(e->e.getAttribute(pageSelector))
-                .collect(Collectors.toList());
 
-        return elements;
+        return null;
         }
-//        List<WebElement> elements= driver.findElements(By.tagName(pageSelector));
-//        for (WebElement info: elements) {
-//            if(info.equals(null)){
-//                System.out.println("potter we have a problem");
-//            }
-//            System.out.println(info.toString());
-//        }
 
-//    }
 
-    public static void main(String[] args) {
-        NameChecking nameChecking;
-//        nameChecking.checkingName("http://www.rimi.ee", "piim");
-    }
 }